絵文字通知機能を実装

+++ src/misskey_api/endpoint/admin/show_moderation_logs.rs
@@ -6,12 +6,12 @@ use super::super::super::model::ModerationLog;
 #[derive(Debug, Serialize)]
 struct Param {
     i : String,
-    limit : i32,
+    limit : u64,
     r#type : Option<String>,
     userId : Option<String>,
 }
 
-pub async fn show_moderation_logs(host : &str, token : &str, limit : i32, r#type : Option<String>, user_id : Option<String>)-> Result<Vec<ModerationLog>, Box<dyn std::error::Error>> {
+pub async fn show_moderation_logs(host : &str, token : &str, limit : u64, r#type : Option<String>, user_id : Option<String>)-> Result<Vec<ModerationLog>, Box<dyn std::error::Error>> {
     let param: Param = Param
         {
             i : String::from(token),
@@ -27,13 +27,13 @@ pub async fn show_moderation_logs(host : &str, token : &str, limit : i32, r#type
 #[derive(Debug, Serialize)]
 struct BySinceIdParam {
     i: String,
-    limit : i32,
+    limit : u64,
     r#type : Option<String>,
     userId : Option<String>,
     sinceId : String,
 }
 
-pub async fn show_moderation_logs_by_since_id(host : &str, token : &str, limit : i32, r#type : Option<String>, user_id : Option<String>, since_id : &str) -> Result<Vec<ModerationLog>, Box<dyn std::error::Error>> {
+pub async fn show_moderation_logs_by_since_id(host : &str, token : &str, limit : u64, r#type : Option<String>, user_id : Option<String>, since_id : &str) -> Result<Vec<ModerationLog>, Box<dyn std::error::Error>> {
     let param: BySinceIdParam = BySinceIdParam
         {
             i : String::from(token),
@@ -50,13 +50,13 @@ pub async fn show_moderation_logs_by_since_id(host : &str, token : &str, limit :
 #[derive(Debug, Serialize)]
 struct ByUntilIdParam {
     i : String,
-    limit : i32,
+    limit : u64,
     r#type : Option<String>,
     userId : Option<String>,
     untilId : String,
 }
 
-pub async fn show_moderation_logs_by_until_id(host : &str, token : &str, limit : i32, r#type : Option<String>, user_id : Option<String>, until_id : &str)-> Result<Vec<ModerationLog>, Box<dyn std::error::Error>> {
+pub async fn show_moderation_logs_by_until_id(host : &str, token : &str, limit : u64, r#type : Option<String>, user_id : Option<String>, until_id : &str)-> Result<Vec<ModerationLog>, Box<dyn std::error::Error>> {
     let param: ByUntilIdParam = ByUntilIdParam
         {
             i : String::from(token),
@@ -73,14 +73,14 @@ pub async fn show_moderation_logs_by_until_id(host : &str, token : &str, limit :
 #[derive(Debug, Serialize)]
 struct BySinceIdUntilIdParam {
     i : String,
-    limit : i32,
+    limit : u64,
     r#type : Option<String>,
     userId : Option<String>,
     untilId : String,
     sinceId : String,
 }
 
-pub async fn show_moderation_logs_by_until_id_since_id(host : &str, token : &str, limit : i32, r#type : Option<String>, user_id : Option<String>, until_id : &str, since_id : &str)-> Result<Vec<ModerationLog>, Box<dyn std::error::Error>> {
+pub async fn show_moderation_logs_by_until_id_since_id(host : &str, token : &str, limit : u64, r#type : Option<String>, user_id : Option<String>, until_id : &str, since_id : &str)-> Result<Vec<ModerationLog>, Box<dyn std::error::Error>> {
     let param: BySinceIdUntilIdParam = BySinceIdUntilIdParam
         {
             i : String::from(token),
